+++ packages/card/ActionIcons.tsx
@@ -20,8 +20,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import * as React from 'react';
+import React from 'react';
 import classnames from 'classnames';
+
+import {CSS_CLASSES} from './constant';
 
 type ChildType = React.ReactElement<React.HTMLProps<HTMLImageElement|HTMLOrSVGElement>>;
 
@@ -34,8 +36,8 @@ const addIconClassToChildren = (children: ChildType | ChildType[]) => {
   return React.Children.map((children as ChildType | ChildType[]), (item) => {
     const className = classnames(
       (item as ChildType).props.className,
-      'mdc-card__action',
-      'mdc-card__action--icon'
+      CSS_CLASSES.ACTION,
+      CSS_CLASSES.ACTION_ICON,
     );
     const props = Object.assign({}, (item as ChildType).props, {className});
     return React.cloneElement((item as ChildType), props);
@@ -45,7 +47,7 @@ const addIconClassToChildren = (children: ChildType | ChildType[]) => {
 const ActionIcons: React.FunctionComponent<ActionIconsProps> = ({
   className = '', children, ...otherProps // eslint-disable-line react/prop-types
 }) => {
-  const classes = classnames('mdc-card__action-icons', className);
+  const classes = classnames(CSS_CLASSES.ACTION_ICONS, className);
   if (!children) return null;
   return (
     <div className={classes} {...otherProps}>