+++ packages/card/PrimaryContent.tsx
@@ -20,11 +20,13 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import * as React from 'react';
+import React from 'react';
 import classnames from 'classnames';
-import * as Ripple from '@material/react-ripple';
+import {withRipple, InjectedProps} from '@material/react-ripple';
 
-export interface PrimaryContentBaseProps extends React.HTMLProps<HTMLDivElement>, Ripple.InjectedProps<HTMLDivElement>{
+import {CSS_CLASSES} from './constant';
+
+export interface PrimaryContentBaseProps extends React.HTMLProps<HTMLDivElement>, InjectedProps<HTMLDivElement>{
   className: string;
   unbounded?: boolean;
 };
@@ -38,7 +40,7 @@ export const PrimaryContentBase: React.FunctionComponent<PrimaryContentBaseProps
   /* eslint-enable react/prop-types */
   ...otherProps
 }) => {
-  const classes = classnames('mdc-card__primary-action', className);
+  const classes = classnames(CSS_CLASSES.PRIMARY_ACTION, className);
 
   return (
     <div className={classes} ref={initRipple} {...otherProps}>
@@ -47,4 +49,4 @@ export const PrimaryContentBase: React.FunctionComponent<PrimaryContentBaseProps
   );
 };
 
-export default Ripple.withRipple<PrimaryContentBaseProps, HTMLDivElement>(PrimaryContentBase);
+export default withRipple<PrimaryContentBaseProps, HTMLDivElement>(PrimaryContentBase);