+++ packages/floating-label/index.tsx
@@ -20,10 +20,11 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import * as React from 'react';
+import React from 'react';
 import classnames from 'classnames';
-// @ts-ignore no mdc .d.ts
-import {MDCFloatingLabelFoundation} from '@material/floating-label/dist/mdc.floatingLabel';
+import {MDCFloatingLabelFoundation} from '@material/floating-label/foundation';
+import {MDCFloatingLabelAdapter} from '@material/floating-label/adapter';
+import {cssClasses} from '@material/floating-label/constants';
 
 export interface FloatingLabelProps extends React.LabelHTMLAttributes<HTMLLabelElement> {
   className?: string;
@@ -39,8 +40,8 @@ export default class FloatingLabel extends React.Component<
   FloatingLabelProps,
   FloatingLabelState
   > {
-  foundation_?: MDCFloatingLabelFoundation;
-  labelElement_: React.RefObject<HTMLLabelElement> = React.createRef();
+  foundation!: MDCFloatingLabelFoundation;
+  labelElement: React.RefObject<HTMLLabelElement> = React.createRef();
 
   static defaultProps: Partial<FloatingLabelProps> = {
     className: '',
@@ -55,12 +56,12 @@ export default class FloatingLabel extends React.Component<
     this.initializeFoundation();
     this.handleWidthChange();
     if (this.props.float) {
-      this.foundation_.float(true);
+      this.foundation.float(true);
     }
   }
 
   componentWillUnmount() {
-    this.foundation_.destroy();
+    this.foundation.destroy();
   }
 
   componentDidUpdate(prevProps: FloatingLabelProps) {
@@ -68,32 +69,41 @@ export default class FloatingLabel extends React.Component<
       this.handleWidthChange();
     }
     if (this.props.float !== prevProps.float) {
-      this.foundation_.float(this.props.float);
+      this.foundation.float(this.props.float!);
     }
   }
 
   initializeFoundation = () => {
-    this.foundation_ = new MDCFloatingLabelFoundation(this.adapter);
-    this.foundation_.init();
+    this.foundation = new MDCFloatingLabelFoundation(this.adapter);
+    this.foundation.init();
   };
 
   get classes() {
     const {classList} = this.state;
     const {className} = this.props;
-    return classnames('mdc-floating-label', Array.from(classList), className);
+    return classnames(cssClasses.ROOT, Array.from(classList), className);
   }
 
-  get adapter() {
+  get adapter(): MDCFloatingLabelAdapter {
     return {
       addClass: (className: string) =>
         this.setState({classList: this.state.classList.add(className)}),
       removeClass: this.removeClassFromClassList,
+      // the adapter methods below are effectively useless since React
+      // handles events and width differently
+      registerInteractionHandler: () => undefined,
+      deregisterInteractionHandler: () => undefined,
+      // Always returns 0 beacuse MDC Web component does
+      // only proxies to foundation.getWidth.
+      // MDC React instead passes it from the text-field
+      // component to floating-label component.
+      getWidth: () => 0,
     };
   }
 
   // must be called via ref
   shake = () => {
-    this.foundation_.shake(true);
+    this.foundation.shake(true);
   };
 
   removeClassFromClassList = (className: string) => {
@@ -103,15 +113,13 @@ export default class FloatingLabel extends React.Component<
   };
 
   handleWidthChange = () => {
-    const {handleWidthChange} = this.props;
-    if (handleWidthChange && this.labelElement_.current) {
-      handleWidthChange(this.labelElement_.current.offsetWidth);
+    if (this.props.handleWidthChange && this.labelElement.current) {
+      this.props.handleWidthChange(this.labelElement.current.offsetWidth);
     }
   };
 
   onShakeEnd = () => {
-    const {LABEL_SHAKE} = MDCFloatingLabelFoundation.cssClasses;
-    this.removeClassFromClassList(LABEL_SHAKE);
+    this.removeClassFromClassList(cssClasses.LABEL_SHAKE);
   };
 
   render() {
@@ -126,7 +134,7 @@ export default class FloatingLabel extends React.Component<
     return (
       <label
         className={this.classes}
-        ref={this.labelElement_}
+        ref={this.labelElement}
         onAnimationEnd={this.onShakeEnd}
         {...otherProps}
       >