+++ packages/layout-grid/Row.tsx
@@ -20,8 +20,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-import * as React from 'react';
+import React from 'react';
 import classnames from 'classnames';
+
+import {CSS_CLASSES} from './constant';
 
 export interface RowProps<T> extends React.HTMLProps<T> {
   className?: string,
@@ -36,7 +38,7 @@ const Row: <T extends {} = HTMLDivElement>(props: RowProps<T>) => React.ReactEle
   ...otherProps
   /* eslint-enable react/prop-types */
 }) => {
-  const classes = classnames('mdc-layout-grid__inner', className);
+  const classes = classnames(CSS_CLASSES.INNER, className);
 
   return (
     // https://github.com/Microsoft/TypeScript/issues/28892