+++ test/unit/top-app-bar/index.test.tsx
@@ -1,7 +1,7 @@
-import * as React from 'react';
+import React from 'react';
 import {assert} from 'chai';
 import {mount, shallow} from 'enzyme';
-import * as td from 'testdouble';
+import td from 'testdouble';
 import TopAppBar, {
   TopAppBarIcon,
   TopAppBarRow,
@@ -43,7 +43,9 @@ class TopAppBarWithScroll extends React.Component<ScrollProps, ScrollState> {
             <TopAppBarSection><TopAppBarTitle>Scroll</TopAppBarTitle></TopAppBarSection>
           </TopAppBarRow>
         </TopAppBar>
-        <div ref={this.state.scrollRef}>Scroll Target</div>
+        <div ref={this.state.scrollRef} style={{height: '1000px', overflow: 'auto'}}>
+          <div style={{height: '3000px'}}>Scroll Target</div>
+        </div>
       </div>
     );
   }
@@ -158,7 +160,7 @@ test('Updating scrollTarget prop will call foundation method destroyScrollHandle
   const wrapper = mount<TopAppBarWithScroll>(<TopAppBarWithScroll/>);
   const topAppBar: TopAppBar = coerceForTesting<TopAppBar>(wrapper.find('TopAppBar').instance());
   const foundation = topAppBar.foundation;
-  foundation.destroyScrollHandler = td.func();
+  foundation.destroyScrollHandler = td.func<() => void>();
   wrapper.instance().withRef();
 
   td.verify(foundation.destroyScrollHandler(), {times: 1});
@@ -168,7 +170,7 @@ test('Updating scrollTarget prop will call foundation method initScrollHandler',
   const wrapper = mount<TopAppBarWithScroll>(<TopAppBarWithScroll/>);
   const topAppBar: TopAppBar = coerceForTesting<TopAppBar>(wrapper.find('TopAppBar').instance());
   const foundation = topAppBar.foundation;
-  foundation.initScrollHandler = td.func();
+  foundation.initScrollHandler = td.func<() => void>();
   wrapper.instance().withRef();
 
   td.verify(foundation.initScrollHandler(), {times: 1});
@@ -245,6 +247,60 @@ test(
   }
 );
 
+test('#adapter.getViewportScrollY returns same value with scrollTarget.scrollTop', () => {
+  const div = document.createElement('div');
+  document.body.appendChild(div);
+
+  const wrapper = mount<TopAppBarWithScroll>(<TopAppBarWithScroll withRef={true} />, {attachTo: div});
+  const topAppBar: TopAppBar = coerceForTesting<TopAppBar>(wrapper.find('TopAppBar').instance());
+  const scrollTarget = topAppBar.state.scrollTarget!.current!;
+  assert.equal(scrollTarget.scrollTop, topAppBar.adapter.getViewportScrollY());
+
+  // https://github.com/material-components/material-components-web-react/pull/771
+  // We commented out the code below because linux puppeteer version doesn't scroll.
+  // const scrollY = 150;
+  // scrollTarget.scrollTo(0, scrollY);
+  // assert.equal(scrollY, topAppBar.adapter.getViewportScrollY());
+
+  document.body.removeChild(div);
+});
+
+test('#adapter.getViewportScrollY test for changing scrollTarget', () => {
+  const div = document.createElement('div');
+  document.body.appendChild(div);
+
+  const wrapper = mount<TopAppBarWithScroll>(<TopAppBarWithScroll withRef={false} />, {attachTo: div});
+  const topAppBar: TopAppBar = coerceForTesting<TopAppBar>(wrapper.find('TopAppBar').instance());
+  assert.equal(window.pageYOffset, topAppBar.adapter.getViewportScrollY());
+  wrapper.setState({withRef: true});
+
+  const scrollTarget = topAppBar.state.scrollTarget!.current!;
+  assert.equal(scrollTarget.scrollTop, topAppBar.adapter.getViewportScrollY());
+  document.body.removeChild(div);
+});
+
+test('#adapter.registerResizeHandler triggers handler on window resize', () => {
+  const wrapper = shallow<TopAppBar>(<TopAppBar />);
+  const testHandler = coerceForTesting<EventListener>(td.func());
+  wrapper.instance().adapter.registerResizeHandler(testHandler);
+  const event = new Event('resize');
+  window.dispatchEvent(event);
+  td.verify(testHandler(event), {times: 1});
+});
+
+test(
+  '#adapter.registerResizeHandler does not trigger handler ' +
+    'after deregistering scroll handler on window',
+  () => {
+    const wrapper = shallow<TopAppBar>(<TopAppBar />);
+    const testHandler = coerceForTesting<EventListener>(td.func());
+    wrapper.instance().adapter.registerResizeHandler(testHandler);
+    const event = new Event('resize');
+    wrapper.instance().adapter.deregisterResizeHandler(testHandler);
+    window.dispatchEvent(event);
+    td.verify(testHandler(event), {times: 0});
+  }
+);
 
 test('#adapter.getTotalActionItems returns one with one actionItem passed', () => {
   const wrapper = mount<TopAppBar>(
@@ -289,7 +345,7 @@ test('#adapter.getTopAppBarHeight should return clientHeight', () => {
   const div = document.createElement('div');
   // needs to be attached to real DOM to get width
   // https://github.com/airbnb/enzyme/issues/1525
-  document.body.append(div);
+  document.body.appendChild(div);
   const options = {attachTo: div};
   const wrapper = mount<TopAppBar>(
     <TopAppBar>
@@ -300,7 +356,7 @@ test('#adapter.getTopAppBarHeight should return clientHeight', () => {
   const topAppBarHeight = wrapper.instance().adapter.getTopAppBarHeight();
   const realDOMHeight = wrapper.getDOMNode().clientHeight;
   assert.equal(topAppBarHeight, realDOMHeight);
-  div.remove();
+  document.body.removeChild(div);
 });
 
 test('when changes from short to fixed the foundation changes', () => {
@@ -413,7 +469,7 @@ test(
 test('#componentWillUnmount destroys foundation', () => {
   const wrapper = shallow<TopAppBar>(<TopAppBar />);
   const foundation = wrapper.instance().foundation;
-  foundation.destroy = td.func();
+  foundation.destroy = td.func<() => void>();
   wrapper.unmount();
   td.verify(foundation.destroy());
 });